extract menu FirehoseResourceLink from vm
